+++ src/main/java/org/trebol/jpa/services/helpers/ProductCategoryTreeResolverImpl.java
@@ -2,12 +2,11 @@ package org.trebol.jpa.services.helpers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.trebol.config.OperationProperties;
 import org.trebol.jpa.entities.ProductCategory;
 import org.trebol.jpa.repositories.IProductsCategoriesJpaRepository;
 import org.trebol.jpa.services.IProductCategoryTreeResolver;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Positive;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -18,23 +17,19 @@ public class ProductCategoryTreeResolverImpl
   implements IProductCategoryTreeResolver {
 
   private final IProductsCategoriesJpaRepository repository;
-  private static final int MAX_RECURSION = 20; // TODO refactor to properties files
+  private final OperationProperties operationProperties;
 
   @Autowired
-  public ProductCategoryTreeResolverImpl(IProductsCategoriesJpaRepository repository) {
+  public ProductCategoryTreeResolverImpl(
+          IProductsCategoriesJpaRepository repository,
+          OperationProperties operationProperties) {
     this.repository = repository;
+    this.operationProperties = operationProperties;
   }
 
   @Override
   public List<ProductCategory> getBranchesFromRoot(ProductCategory rootBranch) {
-    return this.getBranchesFromRoot(rootBranch, MAX_RECURSION);
-  }
-
-  @Override
-  public List<ProductCategory> getBranchesFromRoot(ProductCategory rootBranch, @Positive @Max(MAX_RECURSION) int depth) {
-    if (depth <= 0) {
-      return List.of();
-    }
+    int depth = operationProperties.getMaxCategoryFetchingRecursionDepth();
     List<ProductCategory> immediateDescendants = repository.findByParent(rootBranch);
     List<ProductCategory> allBranches = new ArrayList<>(immediateDescendants);
     for (ProductCategory b : immediateDescendants) {
@@ -45,10 +40,11 @@ public class ProductCategoryTreeResolverImpl
 
   @Override
   public List<Long> getBranchIdsFromRootId(Long rootId) {
+    int depth = operationProperties.getMaxCategoryFetchingRecursionDepth();
     List<Long> immediateDescendantIds = repository.findIdsByParentId(rootId);
     List<Long> allBranches = new ArrayList<>(immediateDescendantIds);
     for (Long bId : immediateDescendantIds) {
-      this.recursivelyAddBranchIds(allBranches, bId, Integer.MAX_VALUE);
+      this.recursivelyAddBranchIds(allBranches, bId, depth);
     }
     return allBranches;
   }